+++ ai/src/greedy.rs
@@ -126,7 +126,7 @@ pub fn play_best_move(state: &mut GameState, weights: &Weights) {
     // The piece is already at its resting row when HardDrop applies (0
     // simulated drop distance from there), so credit the hard-drop score the
     // piece would have earned falling from its original spawn row.
-    state.score += drop_distance * engine::scoring::HARD_DROP_POINTS_PER_CELL;
+    state.score += drop_distance as u64 * engine::scoring::HARD_DROP_POINTS_PER_CELL;
 }
 
 /// A cross-mode placement: which well plus the same rotation/column/row a
@@ -184,7 +184,7 @@ pub fn play_best_cross_move(cross: &mut CrossGame, weights: &Weights) {
     let drop_distance = (placement.row - active.row).max(0) as u32;
     cross.force_active_placement(placement.rotation, placement.row, placement.column);
     cross.apply(Action::HardDrop);
-    cross.wells[placement.arm.index()].score += drop_distance * engine::scoring::HARD_DROP_POINTS_PER_CELL;
+    cross.wells[placement.arm.index()].score += drop_distance as u64 * engine::scoring::HARD_DROP_POINTS_PER_CELL;
 }
 
 #[cfg(test)]
